executes a function given as a parameter

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,3 @@
-#include "function_pointers.h"
 #include <stddef.h>
 
 /**
@@ -11,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array && action)
 		for (i = 0; i < size; i++)
